+++ bin/cache-instagram.mjs
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 import { param_r_ } from '@ctx-core/cli-args'
-import { put_webdriver_scrape_cache } from '../lib/index.js'
+import { webdriver__scrape__cache__put } from '../lib/index.js'
 const { reload, help } = param_r_({
 	reload: '-r, --reload',
 	help: '-h, --help',
@@ -15,4 +15,4 @@ Options:
 -r, --reload  Reload full cache
 	`)
 }
-put_webdriver_scrape_cache(options)
+webdriver__scrape__cache__put(options)