+++ src/created_time_/index.d.ts
@@ -1,0 +1,2 @@
+import type { Medium } from '@ctx-core/instagram'
+export declare function created_time_(medium:Medium):number