+++ src/dynamodb/index.d.ts
@@ -1,0 +1,6 @@
+import type { AttributeValue } from '@aws-sdk/client-dynamodb'
+export declare function put_all_media():Promise<void[]>;
+export declare function _Item(pathname:string):Promise<dynamodb_Item_I>;
+export interface dynamodb_Item_I {
+	[key:string]:AttributeValue;
+}