+++ src/medium_href_a_/index.d.ts
@@ -1,0 +1,6 @@
+export declare function medium_href_a_(opts?:medium_href_a__opts_T):Promise<string[]>
+export { medium_href_a_ as _arr__href__medium }
+export interface medium_href_a__opts_T {
+	INSTAGRAM_NAME?:string
+	reload?:boolean
+}