+++ src/medium_href_a_/index.js
@@ -1,25 +1,30 @@
 import webdriver from 'selenium-webdriver'
 import { medium_pathname_a_ } from '@ctx-core/instagram_cache'
 import { sleep } from '@ctx-core/sleep'
-export async function medium_href_a_(opts:medium_href_a__opts_T = {}):Promise<string[]> {
+/**
+ * @param opts{import('./index.d.ts').medium_href_a__opts_T}
+ * @returns {Promise<string[]>}
+ */
+export async function medium_href_a_(opts = {}) {
 	const {
 		INSTAGRAM_NAME = process.env.INSTAGRAM_NAME,
 		reload,
 	} = opts
-	const current_medium_pathname_a =
-		reload
-		? []
-		: await medium_pathname_a_()
+	const current_medium_pathname_a = reload ? [] : await medium_pathname_a_()
 	const current_media_set = new Set(current_medium_pathname_a)
 	const chrome_Capabilities = webdriver.Capabilities.chrome()
-	chrome_Capabilities.set('chromeOptions', { args: ['--headless'] })
-	const driver = new webdriver.Builder()
-		.forBrowser('chrome')
-		.withCapabilities(chrome_Capabilities)
-		.build()
+	chrome_Capabilities.set('chromeOptions', {
+		args: [
+			'--headless'
+		]
+	})
+	const driver = new webdriver.Builder().forBrowser('chrome').withCapabilities(chrome_Capabilities).build()
 	await driver.get(`https://www.instagram.com/${INSTAGRAM_NAME}/`)
 	let medium_href_a = current_medium_pathname_a
-	let iteration = { href_a_length: 0, iteration_count: 0 }
+	let iteration = {
+		href_a_length: 0,
+		iteration_count: 0
+	}
 	do {
 		await driver.executeScript('window.scrollBy(0, window.innerHeight)')
 		const href_a = JSON.parse(await driver.executeScript(`
@@ -40,7 +45,12 @@ return JSON.stringify(
 	)
 )
 		`.trim()))
-		medium_href_a = [...new Set([...href_a, ...medium_href_a])]
+		medium_href_a = [
+			...new Set([
+				...href_a,
+				...medium_href_a
+			])
+		]
 		if (iteration.href_a_length != medium_href_a.length) {
 			iteration = {
 				href_a_length: medium_href_a.length,
@@ -51,21 +61,17 @@ return JSON.stringify(
 		}
 		if (current_set_any_(href_a)) break
 		await sleep(500)
-		console.debug({ 'medium_href_a.length': medium_href_a.length })
+		console.debug({
+			'medium_href_a.length': medium_href_a.length
+		})
 	} while (iteration.iteration_count < 10)
 	await driver.quit()
 	return medium_href_a
-	function current_set_any_(href_a:string[]) {
+	function current_set_any_(href_a) {
 		for (let i = 0; i < href_a.length; i++) {
 			if (current_media_set.has(href_a[i])) return true
 		}
 		return false
 	}
 }
-export interface medium_href_a__opts_T {
-	INSTAGRAM_NAME?:string
-	reload?:boolean
-}
-export {
-	medium_href_a_ as _arr__href__medium
-}
+export { medium_href_a_ as _arr__href__medium }