+++ src/s3__medium_pathname_a__put/index.d.ts
@@ -1,0 +1,9 @@
+import type { MetadataBearer } from '@aws-sdk/types'
+import type { medium_pathname_a_opts_T } from '@ctx-core/instagram_cache'
+export declare function s3__medium_pathname_a__put(
+	medium_pathname_a:string[], opts?:medium_pathname_a_opts_T
+):Promise<MetadataBearer>
+export {
+	s3__medium_pathname_a__put as s3_put_medium_pathname_a,
+	s3__medium_pathname_a__put as put__arr__pathname__medium,
+}