+++ src/webdriver__scrape__cache__put/index.d.ts
@@ -1,0 +1,7 @@
+import type { medium_href_a__opts_T } from '../medium_href_a_'
+export declare function webdriver__scrape__cache__put(
+	options?:medium_href_a__opts_T
+):Promise<void>
+export {
+	webdriver__scrape__cache__put as put_webdriver_scrape_cache,
+}