+++ CHANGELOG.md
@@ -1,5 +1,18 @@
 # @ctx-core/instagram_cache-node
 
+## 0.1.0
+
+### Minor Changes
+
+- - webdriver**scrape**cache\_\_put: aliased by put_webdriver_scrape_cache
+- extract _.js & _.d.ts into directories
+- - s3**medium_pathname_a**put: aliased by s3_put_medium_pathname_a
+
+### Patch Changes
+
+- Updated dependencies
+  - @ctx-core/instagram_cache@14.0.0
+
 ## 0.0.249
 
 ### Patch Changes